[REF] move sign method in customer service
