+++ shopinvader/tests/__init__.py
@@ -6,7 +6,6 @@ from . import test_backend
 from . import test_cart
 from . import test_cart_item
 from . import test_address
-from . import test_sign
 from . import test_product
 from . import test_sale
 from . import test_shopinvader_variant_binding_wizard
